+++ apps/authoring/src/renderer/components/modals/index.tsx
@@ -1,2 +1,1 @@
 export * from './projects/comp-modal-project-explorer';
-export * from '../../pages/editor/elements/modals/templates/comp-modal-template-explorer';
