+++ apps/authoring/src/renderer/pages/editor/elements/index.ts
@@ -2,3 +2,4 @@ export * from './header';
 export * from './pane-details';
 export * from './right-pane';
 export * from './canvas';
+export * from './modals';
