+++ apps/authoring/src/renderer/pages/editor/elements/modals/template-explorer/editor-modal-template-explorer.module.scss
