+++ apps/authoring/src/renderer/pages/editor/elements/modals/template-explorer/elements/index.ts
@@ -1,0 +1,1 @@
+export * from './template-explorer-body';
