+++ apps/authoring/src/renderer/pages/editor/elements/modals/template-explorer/elements/template-explorer-body.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
-import { Modal, Card, CardBody, Icon, CardTitle, CardHeader } from '@owlui/lib';
-import { Templates } from '../../../../../models';
-import * as styles from './comp-modal-template.module.scss';
-import { updateActiveSlide } from '../../../page-editor-hooks';
+import { Card, CardBody, Icon } from '@owlui/lib';
+import { Templates } from '../../../../../../models';
+import * as styles from '../editor-modal-template-explorer.module.scss';
+import { updateActiveSlide } from '../../../../page-editor-hooks';
 
-const ExplorerTemplateBody = () => {
+export const Body = () => {
   const [isInit, setInit] = useState(false);
   const [templateList, setTemplateList] = useState([]);
 
@@ -24,14 +24,8 @@ const ExplorerTemplateBody = () => {
     });
   }, [isInit]);
 
-  console.log('styles', styles);
-
-  // const updateSlideTemplate = () => {};
   return (
-    <div
-      className={styles.scrowlTemplateBrowser}
-      // style={{ display: 'flex', justifyContent: 'space-evenly' }}
-    >
+    <div className={styles.scrowlTemplateBrowser}>
       {!isInit ? (
         <div>Loading...</div>
       ) : (
@@ -45,12 +39,6 @@ const ExplorerTemplateBody = () => {
               <button
                 onClick={() => updateActiveSlide(item)}
                 className={styles.scrowlTemplateBrowserItem}
-                // style={{
-                //   border: 'none',
-                //   background: 'none',
-                //   boxShadow: 'var(--owl-box-shadow)',
-                //   padding: '0',
-                // }}
                 key={idx}
               >
                 <Card style={{ width: '100%' }}>
@@ -61,14 +49,6 @@ const ExplorerTemplateBody = () => {
                       alt="template-placeholder"
                     />
                   </div>
-                  {/* <CardHeader
-                    style={{
-                      backgroundImage:
-                        'https://eebos.github.io/scrowl_mockup/img/template_0.svg',
-                    }}
-                  >
-                    
-                  </CardHeader> */}
                   <CardBody>
                     <div>{item.manifest.meta.name}</div>
                   </CardBody>
@@ -82,29 +62,6 @@ const ExplorerTemplateBody = () => {
   );
 };
 
-export const ModalExplorerTemplates = () => {
-  const show = Templates.useExplorer();
-  const header = {
-    bsProps: {
-      closeButton: true,
-      closeLabel: 'Close',
-    },
-    content: <>Template Browser</>,
-  };
-  const body = {
-    content: <ExplorerTemplateBody />,
-  };
-
-  return (
-    <Modal
-      show={show}
-      onHide={Templates.closeExplorer}
-      header={header}
-      body={body}
-    />
-  );
-};
-
 export default {
-  ModalExplorerTemplates,
+  Body,
 };
