+++ apps/authoring/src/renderer/pages/editor/elements/modals/template-explorer/index.ts
@@ -1,0 +1,1 @@
+export * from './editor-modal-template-explorer';
