+++ apps/authoring/src/renderer/pages/editor/page-editor.tsx
@@ -1,8 +1,13 @@
 import React from 'react';
 import * as styles from './page-editor.module.scss';
-import { Header, PaneDetails, RightPane, Canvas } from './elements';
+import {
+  Header,
+  PaneDetails,
+  RightPane,
+  Canvas,
+  TemplateExplorerModal,
+} from './elements';
 import { useInit } from './page-editor-hooks';
-import { ModalExplorerTemplates } from '../../components';
 
 export const PageElement = () => {
   const isInit = useInit();
@@ -25,7 +30,7 @@ export const PageElement = () => {
 
             <nav className="scrowl__footer owlui-navbar fixed-bottom"></nav>
           </main>
-          <ModalExplorerTemplates />
+          <TemplateExplorerModal />
         </>
       )}
     </>
